refactor(billing): move coupon redemption into its own service
Extract coupon/promotion redemption (including the Stripe invoice flow) out of
StripeService into a dedicated CouponRedemptionService. It uses the shared Stripe
client for the invoice flow and records the coupon-claim transaction through
StripeTransactionService.recordCouponClaim, so the transactions table stays
written in one place. StripeService sheds the redemption methods and its
transaction-repository dependency. Behavior-preserving.

Named for the domain: this coupon is redeemed for credits (redeemCoupon), as
distinct from discount coupons that would be applied.
